fix: 참여 정보와 맞는 사람만 취소 가능하도록 검증 로직 추가

+++ src/main/java/kakao/rebit/challenge/service/ChallengeParticipationService.java
@@ -79,8 +79,13 @@ public class ChallengeParticipationService {
     }
 
     @Transactional
-    public void cancelParticipation(Long participantId) {
+    public void cancelParticipation(MemberResponse memberResponse, Long participantId) {
         ChallengeParticipation challengeParticipation = findChallengeParticipationByIdOrThrow(participantId);
+
+        if (!challengeParticipation.getMember().getId().equals(memberResponse.id())) {
+            throw new IllegalArgumentException("참여중인 회원이 아닙니다.");
+        }
+
         challengeParticipationRepository.delete(challengeParticipation);
     }
